Added mongodb connection.
Handling mongodb connection in _utils/mongodb.ts
Added mongodb route in api/mongodb/route.ts to test the connection.

added global.d.ts to declare global variables for mongodb in this case.

Added routes for auth as well, but the file is empty for now.

added the mongodb and the nextauth package to the package.json file.