Visual changes to the Poem page and User page.
In the Poem page, I added a wrapping Paper to be
default for a Poem. This creates a nice polaroid
effect when lookin gat a particular poemn with
the edit, delete, tags, etc around it.

In user I made a better poem title card, made the margins
and paddings consistent. Still uncertain about the colour of
the edit button.