Began planning new poetry view components But then saw an issue with the database poems i was trimming the lines of the poems in the database and therefore the display of the poem was off.
Had to change the cleanup script and then, re-convert
them into my schema and finally re-seed the database.

So, I had to change the seedDatabase.ts file and the
mongodb.ts file to load .env.local when the project
isn't actually running.

Anyway. database is now seeded with the correct data
and will now begin working on the new components, But
want to commit this first.