Added the field pageCount to the Poem Type.
I subsequently made sure a pageCount was initialized to 0 in the PoemForm component.
This is then properly set in the backend when a poem is created or updated.

I abstracted the splitPoemIntoPages from the poemcolumns component into its own file.
This function takes a poem and returns an array of pages, each page being an array of lines.
It can also take an optional flag to return only the pageCount.

I used this function in the edit and upload api routes.

Next steps is making poemcolumns and poempage get their
info from the poem object itself rather than try to calculate it themselves.

But i think poemcolumns will still be splitting the poem into pages itself.