Made a better looking usermenu.
FUnctionaility is not there, its jsut the shell, but upload poem does take to the upload poem page.
Sign in and sign out are not working yet.
And the random poem button is not working yet.

Random poem button is probably going to be a link to the random poem page,
so we can probably just make it a link, but i do want to add a tiny d20
 icon next to it or as a loading icon just for flair.

 the rest isf ine, i might make thinsg bitter in genral, jsut teh whole ui .25x atr some point
 but that's not a priority.